API

Index

QuantumGraining.AbstractDiagramNode — Type

decomp.jl contains all recursive functions that help produce all child diagrams from a given contraction.

QuantumGraining.BVector — Type

BVector{T} <: AbstractVector{T}

A struct representing the up- or down-modes. The poles are stored in the poles field.

Arguments

- freqs: A vector representing the frequencies of the vector.
- poles: A vector of the indices of poles of the vector, if any. Does not include the special mode, if there is one.
- special: A boolean representing whether the vector is the special mode.
- type: A symbol representing the type of the vector. Must be :up or :down.

QuantumGraining.Bubble — Type

Bubble{T1, T2} <: AbstractVector{Tuple{T1, T2}}

A struct representing a bubble, also takes care of the edge case.

Fields

- up: A BVector representing the up-bubble.
- down: A BVector representing the down-bubble.
- shape: A tuple representing the shape of the bubble.
- special: A boolean representing whether the bubble is a special edge mode bubble.

QuantumGraining.ContractionCoefficient — Type

struct ContractionCoefficient

A struct representing a contraction coefficient.

Fields

corrections::Vector{Correction}: Vector of Correction objects.

exponents::Vector{Number}: Vector of exponents.

prefacs::Vector{Number}: Vector of prefactors.

polys::Vector{Vector{Number}}: Vector of polynomials.

Constructors

ContractionCoefficient(exponents, prefacs, polys): Constructs a ContractionCoefficient object with given exponents, prefactors, and polynomials.

ContractionCoefficient(exponents, prefacs): Constructs a ContractionCoefficient object with given exponents and prefactors. Polynomials are set to [1] by default.

QuantumGraining.Correction — Type

struct Correction

A struct representing a correction term.

Fields

prefac: The prefactor of the correction term.

exponent: The exponent of the correction term.

poly: A vector of coefficients for the polynomial term.

order: The order of the polynomial term.

Constructors

Correction(prefac, exponent, poly=Num[1,], order=length(poly)): Constructs a Correction object with the given parameters.

If poly is not provided, it defaults to [1]. If order is not provided, it defaults to the length of poly.

QuantumGraining.Diagram — Type

Diagram{T1, T2} <: AbstractVector{Tuple{T1, T2}}

A struct representing a diagram, also takes care of the edge mode.

Arguments:

bubbles: A vector of bubbles.

freqs: A vector of tuples of BVectors. The first element of the tuple is the up-bubble, the second is the down-bubble.

shape: A vector of tuples representing the shape of the bubbles.

up_poles: A vector of vectors representing the up-poles of the bubbles.

down_poles: A vector of vectors representing the down-poles of the bubbles.

num_poles: An integer representing the number of poles in the diagram.

QuantumGraining.DiagramNode — Type

DiagramNode

The DiagramNode struct is a type for one bubble in the diagram. It is defined as a recursive tree where the elements represent the different decompositions of one node into two.

Fields:

root::Tuple{Int, Int}: The root of the tree, the largest common bubble defining the contraction.

val::Tuple{Int, Int}: The value of the current node, determines how much of the bubble was broken to the left.

rightmost::Tuple{Int, Int}: The value of the rightmost bubble.

right::DiagramNode: Pointer to the next right node of the tree, a diagram with one mode broken to the down-bubble.

left::DiagramNode: Pointer to the next left node of the tree, a diagram with one mode broken to the up-bubble.

QuantumGraining.NullNode — Type

NullNode <: AbstractDiagramNode

A terminating node in the tree, has no value. Here so we can check if node can be decomposed using isa DiagramNode

Fields

root: Root value of the tree.

Base.:* — Method

*(c1::Correction, c2::Correction)

Multiply two Correction objects.

Arguments

c1::Correction: The first Correction object.

c2::Correction: The second Correction object.

Base.:* — Method

*(c::Correction, n::Number)

Multiply a Correction object c by a number n.

Arguments

c::Correction: The Correction object to be multiplied.

n::Number: The number to multiply c by.

Base.:* — Method

*(n::Number, c1::ContractionCoefficient)

Multiply a number n with a ContractionCoefficient object c1.

Arguments

n::Number: The number to be multiplied.

c1::ContractionCoefficient: The ContractionCoefficient object to be multiplied.

Base.:* — Method

*(n::Number, c::Correction)

Multiply a number n with a Correction object c.

Arguments

n::Number: The number to be multiplied.

c::Correction: The Correction object to be multiplied.

Base.:+ — Method

+(c1::ContractionCoefficient, c2::ContractionCoefficient)

Addition operator for ContractionCoefficient objects. This function performs element-wise addition of two ContractionCoefficient objects, c1 and c2. It creates a new ContractionCoefficient object by adding the corresponding elements of c1 and c2. The resulting object is then simplified by merging duplicate exponents and simplifying the contraction.

Arguments

c1::ContractionCoefficient: The first ContractionCoefficient object.

c2::ContractionCoefficient: The second ContractionCoefficient object.

Returns

ContractionCoefficient: The result of the addition operation.

Base.:+ — Method

+(c1::ContractionCoefficient, c2::Correction)

Add a Correction object to a ContractionCoefficient object.

Arguments

c1::ContractionCoefficient: The ContractionCoefficient object to add to.

c2::Correction: The Correction object to add.

Returns

A new ContractionCoefficient object with the Correction object added.

Base.:+ — Method

+(c1::Correction, c2::Correction)

Addition operator for Correction objects. This function adds two Correction objects c1 and c2 and returns a new ContractionCoefficient object.

Arguments

c1::Correction: The first Correction object.

c2::Correction: The second Correction object.

Base.:== — Method

==(c1::ContractionCoefficient, c2::ContractionCoefficient)

Check if two ContractionCoefficient objects are equal by comparing their corrections arrays.

Arguments

c1::ContractionCoefficient: The first ContractionCoefficient object.

c2::ContractionCoefficient: The second ContractionCoefficient object.

Returns

true if the corrections arrays of c1 and c2 are equal, false otherwise.

Base.:== — Method

==(c1::Correction, c2::Correction)

Check if two Correction objects are equal by comparing their order, exponent, and poly properties.

Arguments

c1::Correction: The first Correction object.

c2::Correction: The second Correction object.

Returns

true if the order, exponent, and poly properties of c1 and c2 are equal, false otherwise.

QuantumGraining.DVec — Method

" DVec(u::Vector; special=false) Creates a BVector with type :down. Assumes the frequency ordering is clockwise

Arguments

- `u::Vector`: Components of the vector.
- `special::Bool=false`: Whether the vector holds the special mode.

QuantumGraining.UVec — Method

" UVec(u::Vector; special=false) Creates a BVector with type :up. Assumes the frequency ordering is clockwise

Arguments

- `u::Vector`: Components of the vector.
- `special::Bool=false`: Whether the vector holds the special mode.

QuantumGraining.contraction_coeff — Method

contraction_coeff(left::Int, right::Int, freqs::Array)

Calculates the coefficient of a whole contraction, given the contraction and input frequencies. Calculates equation (7) in the paper.

Arguments

left: the left-order of the contraction

right: the right-order of the contraction

freqs: array of frequencies to put in each mode

Returns

c: a contraction coefficient struct, symbolic expression for the contraction coeffeicient.

QuantumGraining.contraction_to_function — Method

contraction_to_function(g, ω, ps)

Convert a contraction expression g into a Julia function that depends on the variables t and τ.

Arguments

g: The contraction expression to be converted.

ω: The angular frequency.

ps: Additional parameters.

QuantumGraining.conv — Method

conv(u::Vector{<:Number}, v::Vector{<:Number})

Compute the convolution of two vectors u and v.

Arguments

u::Vector{<:Number}: The first input vector.

v::Vector{<:Number}: The second input vector.

Returns

result::Vector{<:Number}: The resulting vector after performing the convolution.

QuantumGraining.count_modes — Method

count_modes(diagram)

Counts the total number of up-modes and down-modes in a given diagram. Due to the way the diagrams are constructed now, this will be deprecated soon.

QuantumGraining.count_poles — Method

count_poles(s_list::Vector{Int}, stag_list::Vector{Int})

Given two lists of poles, counts the total number of poles by counting the non-empty lists.

Arguments:

s_list: list of upper poles

stag_list: list of lower poles

QuantumGraining.diagram_correction — Method

diagram_correction(ω)

Gives an expression for the effective diagram correction.

Arguments

ω: A list of vectors (μi, νi) with the frequency values for each mode.

For example: For a diagram d = [(3, 2), (2, 1)] we would have ω = [(μ1, ν1), (μ2, ν2)] where μi and νi are vectors containing mode values.

Returns

A Correction struct representing the effective diagram correction.

QuantumGraining.drop_high_freqs — Function

drop_high_freqs(freqs_list::Vector, freqs_subs, cutoff=0.1)

Given a list of frequencies and a list of substitutions, returns only the low frequencies.

Arguments

freqs_list::Vector: A list of frequencies.

freqs_subs: A list of substitutions.

cutoff=0.1: The cutoff value for determining low frequencies.

Returns

rwa: A list of indices corresponding to the low frequencies.

freqs_low: A list of low frequencies.

QuantumGraining.drop_high_freqs — Method

drop_high_freqs(gs_dict, freqs_dict, freqs_vals; cutoff=0.1)

Given dictionaries holding the frequencies and couplings, drops all high-frequency contributions.

Arguments

gs_dict: A dictionary holding the coupling strengths.

freqs_dict: A dictionary holding the frequencies.

freqs_vals: A dictionary of substitutions for the frequencies.

cutoff=0.1: The cutoff value for determining high frequencies.

Returns

gs_dict: The updated dictionary of coupling strengths after dropping high-frequency contributions.

freqs_dict: The updated dictionary of frequencies after dropping high-frequency contributions.

QuantumGraining.effective_dissipator — Method

effective_dissipator(h::Vector, gs::Vector, Ω::Vector, k::Int; as_dict=true)

Compute the effective dissipator for a given set of parameters.

Arguments

h::Vector: A vector of operators representing the raw Hamiltonian.

gs::Vector: A vector of coupling strengths.

Ω::Vector: A vector of frequencies.

k::Int: The truncation order.

as_dict::Bool: (optional) If true, returns the dissipator as a dictionary of operators and their corresponding dissipator terms. If false, returns the dissipator as separate vectors for operators, dissipator terms, and frequencies. Default is true.

Returns

If as_dict is true, returns a tuple (γs_dict, ωs_dict) where γs_dict is a dictionary of operators and their corresponding dissipator terms, and ωs_dict is a dictionary of operators and their corresponding frequencies.

If as_dict is false, returns a tuple (ops_eff, γs_eff, ωs_eff) where ops_eff is a vector of operators representing the effective dissipator, γs_eff is a vector of dissipator terms, and ωs_eff is a vector of frequencies.

QuantumGraining.effective_dissipator_term — Method

effective_dissipator_term(h::Vector, gs::Vector, Ω::Vector, k::Int)

Compute the effective dissipator term for a given set of parameters.

Arguments

h::Vector: Vector of coefficients for the Hamiltonian terms.

gs::Vector: Vector of coefficients for the jump operators.

Ω::Vector: Vector of coefficients for the frequencies.

k::Int: Number of terms in the Hamiltonian.

Returns

J_list::Vector: Vector of tuples representing the jump operators.

γ_list::Vector: Vector of complex numbers representing the dissipation coefficients.

ω_list::Vector: Vector of sums of frequencies.

QuantumGraining.effective_hamiltonian — Method

effective_hamiltonian(h::Vector, gs::Vector, Ω::Vector, k::Int; as_dict=false, remove_constants=true)

Compute the effective Hamiltonian for a given system.

Arguments

h::Vector: A vector of Hamiltonian terms.

gs::Vector: A vector of coupling strengths.

Ω::Vector: A vector of frequencies.

k::Int: The number of terms to consider.

as_dict::Bool=false: If true, the output will be returned as a dictionary of operators and frequencies.

remove_constants::Bool=true: If true, remove constant terms from the output.

Returns

If as_dict is true, returns a tuple (gs_eff, ωs_eff) where gs_eff is a dictionary of operators and ωs_eff is a dictionary of frequencies.

If as_dict is false, returns a tuple (unique_hs, unique_gs, ωs_eff) where unique_hs is a vector of unique operators, unique_gs is a vector of unique coupling strengths, and ωs_eff is a vector of frequencies.

QuantumGraining.effective_hamiltonian_term — Method

effective_hamiltonian(h::Vector, g::Vector{Number}, Ω::Vector{Number}, k::Int)

Given a truncation order k, a list of frequencies Ω, couplings g, and operators h representing the raw Hamiltonian, this function returns new frequencies, coupling strengths, and operators representing the Hamiltonian to order k.

Details

The function calculates the effective Hamiltonian by considering all possible combinations of operators, frequencies, and coupling strengths up to order k. It simplifies the expressions and merges duplicate coupling strengths.

Arguments

h::Vector: A vector of operators representing the raw Hamiltonian.

g::Vector{Number}: A vector of coupling strengths.

Ω::Vector{Number}: A vector of frequencies.

k::Int: The truncation order.

Returns

ops_eff::Vector: A vector of operators representing the effective Hamiltonian to order k.

merged_gs::Vector: A vector of merged coupling strengths.

ωs_eff::Vector: A vector of frequencies representing the effective Hamiltonian to order k.

QuantumGraining.expand_operator — Method

expand_operator(h)

Expand a quantum operator into a list of individual operators and their corresponding coefficients.

Arguments

h: The quantum operator to be expanded.

Returns

ops: A list of individual operators.

facs: A list of corresponding coefficients.

QuantumGraining.expand_operators — Method

expand_operators(hs, gs, ωs)

Go over any sum of operators and breaks it into the constituent operators while preserving the order of the other vectors.

Arguments

hs: Array of operators to be expanded.

gs: Array of coefficients corresponding to the operators.

ωs: Array of weights corresponding to the operators.

Returns

unique_hs: Array of expanded operators.

unique_gs: Array of coefficients corresponding to the expanded operators.

unique_ωs: Array of weights corresponding to the expanded operators.

QuantumGraining.expand_operators — Method

expand_operators(hs)

Expand a list of operators by applying the expand_operator function to each operator.

Arguments

hs: A list of operators to be expanded.

QuantumGraining.extend_correction — Method

extend_correction(c::Correction, poly::Vector{<:Number})

Extend a correction by normalizing and modifying its polynomial representation.

Arguments

c::Correction: The correction to be extended.

poly::Vector{<:Number}: The polynomial representation of the correction.

Returns

Correction: The extended correction.

QuantumGraining.find_all_poles — Method

find_all_poles(d::Vector{Tuple{Vector{T1}, Vector{T2}}}) where {T1, T2}

Finds all poles in vector factorial for frequency list d. Assumes that the first tuple in the list is the special mode.

Arguments

ω = A vector of bubble frequencies [(μ1, ν1), ..., (μ||d||, ν||d||)]

Returns

up_poles: list of indices of poles in upper modes

down_poles: list of indices of poles in lower modes

QuantumGraining.find_all_poles — Method

find_all_poles(d::Vector{Tuple{BVector{T1}, BVector{T2}}}) where {T1, T2}

Find all poles in the given vector of tuples d. Each tuple contains two BVector objects, μ and ν.

Details

The function iterates over each tuple and performs the following steps:

If μ.special is true, it removes the first element from μ.

Reverses the order of elements in μ.

Calls the find_poles function to find the poles in μ and ν.

Appends the resulting poles to up_poles and down_poles respectively.

Arguments

d::Vector{Tuple{BVector{T1}, BVector{T2}}}: A vector of tuples, where each tuple contains two BVector objects.

Returns

A tuple (up_poles, down_poles) where up_poles is a vector of vectors containing the poles found in μ, and down_poles is a vector of vectors containing the poles found in ν.

QuantumGraining.find_integer_solutions — Function

find_integer_solutions(num_vars::Int, target_sum::Int, combination::Vector{Int}=Vector{Int}(), sum_so_far::Int=0)

Function that calculates combinations of k positive integers adding up to m, as detailed in the paper.

Arguments

num_vars::Int: The number of variables in the combination.

target_sum::Int: The target sum of the combination.

combination::Vector{Int}=Vector{Int}(): The current combination of integers.

sum_so_far::Int=0: The sum of the integers in the current combination.

QuantumGraining.find_poles — Method

find_poles(u)

Finds all the factors of the vector factorial that evaluate to 0.

Arguments

- ω = A vector of mode frequencies [(μ_1, ν_1), ..., (μ_||d||, ν_||d||)]

QuantumGraining.gaussian_to_cutoff — Method

gaussian_to_cutoff(gs, freq_vals; cutoff=0.1, keep_small_exponents=true)

Apply a cutoff to Gaussian exponents based on their absolute value.

Arguments

gs: A list of ContractionCoefficient objects representing Gaussian functions.

freq_vals: A dictionary mapping frequency keys to their corresponding values.

cutoff: The cutoff value for the absolute value of the Gaussian exponents. Default is 0.1.

keep_small_exponents: A boolean indicating whether to keep small exponents or set them to zero. Default is true.

QuantumGraining.gaussian_to_cutoff — Method

gaussian_to_cutoff(gs::Dict, ωs::Dict, freq_vals; cutoff=0.1, keep_small_exponents=true)

Apply a cutoff to Gaussian exponents based on their absolute value.

Arguments

gs: A dictionary mapping frequency keys to ContractionCoefficient objects representing Gaussian functions.

ωs: A dictionary mapping frequency keys to their corresponding values.

freq_vals: A dictionary mapping frequency keys to their corresponding values.

cutoff: The cutoff value for the absolute value of the Gaussian exponents. Default is 0.1.

keep_small_exponents: A boolean indicating whether to keep small exponents or set them to zero. Default is true.

QuantumGraining.get_diagrams — Method

get_diagrams(init_diagram::Array)

Uses the above node_decomp() to get all possible diagrams recursively.

Arguments

node::AbstractDiagramNode: The initial diagram including the first level of breakdowns (two bubble diagrams)

Returns

A list of all possible diagrams for a given contractions.

QuantumGraining.group_operators — Method

group_operators(hs, gs, ωs; as_dict=true)

Group operators based on their values and return the grouped operators.

Arguments

hs: An array of operators.

gs: An array of coefficients corresponding to the operators.

ωs: An array of corresponding elements.

as_dict: A boolean indicating whether to return the grouped operators as dictionaries. Default is true.

Returns

If as_dict is true, returns two dictionaries gs_dict and ωs_dict where the keys are the operators and the values are the grouped coefficients and elements respectively.

If as_dict is false, returns three arrays new_hs, new_gs, and new_ωs where new_hs contains the grouped operators, new_gs contains the grouped coefficients, and new_ωs contains the grouped elements.

QuantumGraining.group_operators — Method

group_operators(hs)

Group operators in the given array hs by removing duplicates.

Arguments

hs: An array of operators.

Returns

new_hs: An array of operators with duplicates removed.

QuantumGraining.hamiltonian_function — Method

hamiltonian_function(gs, ωs, h_src, h_tgt, ps; return_contraction_functions=false)

Constructs a Hamiltonian function based on the given parameters for QuantumOptics.jl.

Arguments

gs: A dictionary of contraction functions, where the keys are quantum numbers and the values are the corresponding contraction functions.

ωs: A dictionary of frequencies, where the keys are quantum numbers and the values are the corresponding frequencies.

h_src: An array of source Hamiltonian operators.

h_tgt: An array of target Hamiltonian operators.

ps: A dictionary of parameters, where the keys are parameter names and the values are the corresponding parameter values.

return_contraction_functions: (optional) A boolean indicating whether to return the contraction functions along with the Hamiltonian function. Default is false.

Returns

If return_contraction_functions is false, returns a Hamiltonian function H_func that takes a time t and a state ψ as input and returns the Hamiltonian operator applied to the state.

If return_contraction_functions is true, returns a tuple (H_func, func_gs) where H_func is the Hamiltonian function and func_gs is a dictionary of contraction functions, where the keys are quantum numbers and the values are the corresponding contraction functions.

QuantumGraining.lindblad_function — Method

lindblad_function(gs, Ωs, γs, ωs, h_src, h_tgt, ps)

Still incomplete. Construct a QuantumOptics.jl Lindblad function that represents the Lindblad master equation for a quantum system.

Details

The Lindblad master equation is given by: dρ/dt = -i[H(t, ρ), ρ] + ∑(J * ρ * L - 0.5 * (L * J * ρ + ρ * L * J)) where H(t, ρ) is the time-dependent Hamiltonian, ρ is the density matrix, J is the jump operator, and L is the adjoint of the jump operator.

The function L_func takes two arguments: t (time) and ρ (density matrix), and returns a tuple (H, J, L, rates). H is the time-dependent Hamiltonian, J and L are arrays of jump operators, and rates is an array of associated decay/jump rates.

Arguments

gs: An array of complex numbers representing the coupling strengths between the system and the environment.

Ωs: An array of complex numbers representing the Rabi frequencies of the system.

γs: An array of tuples representing the jump operators and their corresponding decay rates.

ωs: A dictionary mapping jump operators to their corresponding frequencies.

h_src: An array of quantum operators representing the source Hamiltonian.

h_tgt: An array of quantum operators representing the target Hamiltonian.

ps: A dictionary mapping jump operators to their corresponding probabilities.

QuantumGraining.merge_duplicate_exponents — Method

merge_duplicate_exponents(c::ContractionCoefficient)

Merge duplicate exponents in a ContractionCoefficient object.

Details

This function takes a ContractionCoefficient object and merges any duplicate exponents. It iterates over the exponents, prefactors, and polynomials in the ContractionCoefficient object and checks for duplicates. If duplicates are found, the prefactors and polynomials are merged into a single polynomial. The resulting unique exponents, merged prefactors, and merged polynomials are then used to create a new ContractionCoefficient object. Note that the function uses the ordered_sum function to merge the polynomials and the simplify function to simplify the resulting polynomial.

Arguments

c::ContractionCoefficient: The ContractionCoefficient object to merge duplicate exponents.

QuantumGraining.node_decomp! — Method

node_decomp(node::DiagramNode)
node_decomp!(node::AbstractDiagramNode, decomp_list)

Use the DiagramNode structure to give one level of decompositions explicitly using a recursive function. In other words, gives all ways one can break a bubble into two.

Arguments

node::DiagramNode: Node to break down.

Returns

decomp_list::Array{Int}: a list of all nodes in the tree.

QuantumGraining.norm_fac — Method

norm_fac(v::BVector, j::Int, mj::Int)

Computes the normalization factor for a given BVector object v at mode j and mj.

Arguments

v::BVector: The input BVector object.

j::Int: The mode index.

mj::Int: The inner mode index.

Returns

Float64: The normalization factor.

QuantumGraining.ordered_sum — Method

ordered_sum(a, b)

Compute the element-wise sum of two arrays a and b, padding the shorter array with zeros.

Arguments

a: The first input array.

b: The second input array.

Returns

The element-wise sum of a and b, with the shorter array padded with zeros.

QuantumGraining.pad — Method

pad(a::Vector, new_length)

Pad a vector a with zeros to a new length new_length.

Details

The padded vector has length new_length, where the original elements of a are preserved and the remaining elements are filled with zeros.

Arguments

a::Vector: The input vector to be padded.

new_length: The desired length of the padded vector.

QuantumGraining.qc_convert — Method

qc_convert(gs, ωs, cnumbers_dict, t, τ)

Convert a ContractionCoefficient object to an array of QuantumCumulants operators.

Arguments

gs: A dictionary mapping operators to their corresponding contraction coefficients.

ωs: A dictionary mapping operators to their corresponding frequencies.

cnumbers_dict: A dictionary mapping contraction coefficients to their corresponding values.

QuantumGraining.qc_convert — Method

qc_convert(gs::Vector, ops::Vector, freqs::Vector, cnumbers_dict, t, τ)

Convert a ContractionCoefficient object to an array of QuantumCumulants operators.

Arguments

gs::Vector: An array of contraction coefficients.

ops::Vector: An array of operators.

freqs::Vector: An array of frequencies.

cnumbers_dict: A dictionary mapping contraction coefficients to their corresponding values.

QuantumGraining.qnumber_to_qop — Method

qnumber_to_qop(qn::QuantumCumulants.QAdd, op_subs, Id; mul = tensor)

Convert a QuantumCumulants.QAdd object to a QAdd quantum operator.

Arguments

qn::QuantumCumulants.QAdd: The QAdd object representing the quantum number.

op_subs: The operator substitutions to be applied.

Id: The identity operator.

mul: The function used for tensor multiplication. Default is tensor.

QuantumGraining.qnumber_to_qop — Method

qnumber_to_qop(qn::QuantumCumulants.QMul, op_subs, Id; mul = tensor)

Convert a QuantumCumulants.QMul object to a QMul quantum operator.

Arguments

qn::QuantumCumulants.QMul: The QMul object representing the quantum number.

op_subs: The operator substitutions to be applied.

Id: The identity operator.

mul: The function used for tensor multiplication. Default is tensor.

QuantumGraining.qnumber_to_qop — Method

qnumber_to_qop(qn::QuantumCumulants.QSym, op_subs, Id; mul = tensor)

Convert a QuantumCumulants.QSym object to a QSym quantum operator.

Arguments

qn::QuantumCumulants.QSym: The QSym object representing the quantum number.

op_subs: The operator substitutions to be applied.

Id: The identity operator.

mul: The function used for tensor multiplication. Default is tensor.

QuantumGraining.repeated_combinations — Method

repeated_combinations(h::Vector, Ω::Vector, n::Int)

Given a collection of terms and a number n, returns all possible n repetitions of the elements in the collection.

Arguments

h::Vector: A vector of operators to find the combinations of.

Ω::Vector: A corresponding vector of frequencies.

Returns

perm_h::Vector: A vector of combinations of operators.

perm_Ω::Vector: A vector of combinations of frequencies.

QuantumGraining.repeated_combinations — Method

repeated_combinations(h::Vector, g::Vector, Ω::Vector, n::Int)

Generate all possible combinations of elements from the input vectors h, g, and Ω with repetition, forming combinations of length n.

Arguments

h::Vector: Vector of elements for combination from h

g::Vector: Vector of elements for combination from g

Ω::Vector: Vector of elements for combination from Ω

n::Int: Length of combinations to generate

Returns

perm_h::Vector: Vector of combinations of elements from h

perm_g::Vector: Vector of combinations of elements from g

perm_Ω::Vector: Vector of combinations of elements from Ω

QuantumGraining.reshape_sols — Function

reshape_sols(sols, target_sum, num_bubbles, num_indices = 3)

Helper function that reshapes integer combinations from findintegersolutions() into vectors

Arguments

sols: The integer combinations.

target_sum: The target sum of the combinations.

num_bubbles: The number of bubbles.

num_indices::Int=3: The number of indices in the combination.

QuantumGraining.split_freqs — Method

split_freqs(freqs::Array, ububs::Int, dbubs::Int)

It is useful to have the contraction frequencies in a single list, since it allows more easily to go over all possible permutation. However, other functions need the frequencies to be split into up- and down-bubbles. The function split_freqs splits a single frequency array [μ1, μ2, ..., μl, ν1, ν2, ..., νr] into an array ububs = [μ1, μ2, ..., μl] of up-bubbles and an array dbubs = [ν1, ν2, ..., νr] of down-bubbles.

Arguments: freqs - unified array of frequencies, should be of the form [μ1, μ2, ..., μl, ν1, ν2, ..., νr] l - number of up-modes r - number of down-modes

Returns: freqsup - array of only the up-modes freqsdn - array of only the down-modes

QuantumGraining.split_freqs_into_tuples — Method

split_freqs_into_bubbles(freqs, diagram)

Splits an array of frequencies into an array of tuples of frequencies, matching the dimensions of each bubble in diagram.

Arguments

freqs - an array of frequencies in the form ω = [μ1..., μ2..., ..., μl..., ν1..., ν2..., ..., νr]
diagram - dimensions of each bubble, where each tuple is the dimension for the corresponding bubble.

Returns

ω - an array of frequencies in the form [(μ1, ν1), (μ2, ν2), ..., (μl, νl), ([], ν(l+1)), ..., ([], νr)]

QuantumGraining.symbolic_hamiltonian — Method

symbolic_hamiltonian(gs::Dict, Ω::Dict, t, τ)

Constructs a symbolic Hamiltonian for a quantum system.

Arguments

gs::Dict: A dictionary mapping operators to their corresponding coefficients.

Ω::Dict: A dictionary mapping operators to their corresponding frequencies.

t: The time parameter.

τ: The time step parameter.

Returns

A symbolic Hamiltonian for the quantum system.

QuantumGraining.symbolic_hamiltonian — Method

symbolic_hamiltonian(gs::Vector, ops::Vector, Ω::Vector, t, τ)

Constructs a symbolic Hamiltonian for a quantum system.

Arguments

gs::Vector: Vector of symbols representing the coefficients of the Hamiltonian terms.

ops::Vector: Vector of operators corresponding to each Hamiltonian term.

Ω::Vector: Vector of frequencies for each Hamiltonian term.

t: Time parameter.

τ: Symbolic time parameter.

Returns

terms: Vector of symbolic Hamiltonian terms.

QuantumGraining.taylor_coeff — Method

taylor_coeff(n::Int, k::Int)

Compute the Taylor coefficient for a given pair of integers n and k, as detailed in the paper.

Arguments

n::Int: The total degree of the Taylor series.

k::Int: The degree of the term whose coefficient is being computed.

Returns

The value of the Taylor coefficient for the given pair of integers.

QuantumGraining.to_array — Method

to_array(node::DiagramNode)

Given a node, returns the diagram in array form where each entry corresponds to a different bubble.

Arguments

node::DiagramNode: The node to convert to an array.

QuantumGraining.to_symbol — Method

to_symbol(coeff::ContractionCoefficient, τ)

Give a symbolic representation of a ContractionCoefficient object.

Arguments

coeff::ContractionCoefficient: The ContractionCoefficient object to compute the symbol for.

τ: The value of τ.

Returns

Symbolic representation of the ContractionCoefficient object.

QuantumGraining.to_symbol — Method

to_symbol(c::Correction, τ)

Give a symbolic representation of a Correction object.

Arguments

c::Correction: The Correction object to compute the symbol for.

τ: The value of τ.

Returns

Symbolic representation of the Correction object.

QuantumGraining.vec_factorial — Method

vec_factorial(u::BVector; include_poles=false)

Compute the vector factorial of a BVector object u.

Arguments

u::BVector: The input BVector object.

include_poles::Bool=false: Whether to include poles in the computation.

Returns

The vector factorial of u.

Example

u = UVec([a, b, c])
vec_factorial(u) # (a+b+c)*(a+b)*c